Added proper SQLite support [FATAL BUG]
FATAL BUG
For a reason I have yet to discover, as of this commit the mod can't be used on a server. I get a java.lang.nullpointerexception, stating that the org.sqlite.jdbc class can't be found. However, it is a dependency in build.gradle. Please fix and submit a pull request if you have any ideas